requirements.txt
torch>=1.13.0
transformers>=4.30.0
huggingface_hub
datasets>=2.12.0
accelerate>=0.21.0
peft>=0.3.0
evaluate>=0.4.0
mlflow>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
pandas 
numpy
scikit-learn
fastapi
uvicorn
nltk
rouge_score
sentencepiece
streamlit
dagshub
docker
-e .